Move item specific completion tests
