+++ crates/ide_completion/src/tests.rs
@@ -6,6 +6,7 @@
 
 mod item_list;
 mod use_tree;
+mod items;
 
 use hir::{PrefixKind, Semantics};
 use ide_db::{
